add spi support
